refactor: 상태별로 테스트 파일 분리
