+++ src/test/kotlin/blackjack/domain/state/BustTest.kt
@@ -1,0 +1,4 @@
+package blackjack.domain.state
+
+class BustTest {
+}
